back/requirements.txt
requests
google-search-results
python-dotenv
flask
flask_sqlalchemy
psycopg2-binary
prisma